Project refactoring
Reformate code and create packages
