--- src/CharAriDemo.java
@@ -1,15 +1,0 @@
-public class CharAriDemo{
-    public static void main(String args[]){
-        char ch;
-
-        ch = 'X';
-        System.out.println("теперь ch содержит" + ch);
-
-        ch++;
-        System.out.println("теперь ch содержит" + ch);
-
-        ch = 90;
-        System.out.println("теперь ch содержит" + ch);
-    }
-}
-
