--- src/com/veronika/HelloWorld.java
@@ -1,12 +1,0 @@
-package com.veronika;
-
-public class HelloWorld {
-    public static void main(String[] args) {
-        System
-                .out
-                .println(
-                        "Hello world!"
-                )
-        ;
-    }
-}
