+++ src/com/veronika/part_1/Example2.java
@@ -1,4 +1,4 @@
-package com.veronika;
+package com.veronika.part_1;
 
 class Example2 {
      public static void main(String[] args) {
