+++ src/com/veronika/part_1/Example3.java
@@ -1,7 +1,7 @@
-package com.veronika;
+package com.veronika.part_1;
 
 public class Example3 {
-    public static void main(String args []){
+    public static void main(String[] args){
         int var;
         double x;
         var = 10;
