+++ src/com/veronika/part_1/HelloWorld.java
@@ -1,0 +1,7 @@
+package com.veronika.part_1;
+
+public class HelloWorld {
+    public static void main(String[] args) {
+        System.out.println("Hello world!");
+    }
+}
