+++ src/com/veronika/part_1/IfDemo.java
@@ -1,4 +1,4 @@
-package com.veronika;
+package com.veronika.part_1;
 
 public class IfDemo {
     public static void main(String[] args) {
