+++ src/com/veronika/part_2/CharAriDemo.java
@@ -1,0 +1,17 @@
+package com.veronika.part_2;
+
+public class CharAriDemo {
+    public static void main(String[] args) {
+        char ch;
+
+        ch = 'X';
+        System.out.println("теперь ch содержит " + ch);
+
+        ch++;
+        System.out.println("теперь ch содержит " + ch);
+
+        ch = 90;
+        System.out.println("теперь ch содержит " + ch);
+    }
+}
+
